pyproject.toml
[project]
name = "posts-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Максим Малеев",email = "mmaleev1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (==3.2)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "djangorestframework (>=3.14,<4.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (==4.8.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "django-admin-rangefilter (>=0.13.3,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
flake8 = "^7.3.0"
coverage = "^7.10.1"
pytest-cov = "^6.2.1"

